Fix bug that caused irregular submenu sizing.
It's not a perfect solution, but a good band-aid for now. Trouble
was that a sub-menu with a longer first word than the menu header's
first word would appear wider than its enclosing container.

I've been meaning to rewrite these menus as flexboxes for a while
now. This may very well be the straw that breaks my g*ddamn back.